feat: CustomError 생성
