+++ frontend/src/constants/errorStrategy.ts
@@ -1,0 +1,7 @@
+export const ERROR_STRATEGY = {
+  TOAST: "toast",
+  MODAL: "modal",
+  REDIRECT: "redirect",
+  ERROR_BOUNDARY: "errorBoundary",
+  IGNORE: "ignore",
+} as const;
